Fix int truncation of 64-bit values in debug log messages
Several debug and log messages cast uint64_t and uint32_t values to
int before logging, causing truncation for values exceeding 2^31.

This is visible when using hard drive images larger than 4GB, where
the image file size is reported incorrectly in zululog.txt. For
example a 149GB image was logged as 1128095744 bytes (the lower
32 bits of 0x25433D6000).

Replace (int) casts with (uint64_t) for values that represent file
sizes, LBA addresses, sector counts, byte offsets, and capacities
across ide_imagefile.cpp, ide_rigid.cpp, ide_atapi.cpp, and
ide_cdrom.cpp. Small values like blocksize, heads, and track numbers
are left as int since they fit in 32 bits.

The log_raw(uint64_t) overload already exists in ZuluIDE_log.h so
no additional infrastructure changes are needed.

+++ src/ide_cdrom.cpp
@@ -1215,7 +1215,7 @@ bool IDECDROMDevice::doReadSubChannel(bool time, bool subq, uint8_t parameter, u
         uint8_t audiostatus;
         uint32_t lba;
         cdromGetAudioPlaybackStatus(&audiostatus, &lba, false);
-        dbgmsg("------ Get audio playback position: status ", (int)audiostatus, " lba ", (int)lba);
+        dbgmsg("------ Get audio playback position: status ", (int)audiostatus, " lba ", (uint64_t)lba);
 
         // Fetch current track info
         CUETrackInfo trackinfo = getTrackFromLBA(lba);
@@ -1326,7 +1326,7 @@ bool IDECDROMDevice::doReadCD(uint32_t lba, uint32_t length, uint8_t sector_type
             if (seek_back > offset)
             {
                 logmsg("WARNING: Host attempted CD read at sector ", lba, "+", length,
-                       " pregap request ", (int)seek_back, " exceeded available ", (int)offset, " for track ", trackinfo.track_number,
+                       " pregap request ", (int)seek_back, " exceeded available ", (uint64_t)offset, " for track ", trackinfo.track_number,
                        " (possible .cue file issue)");
                 offset = 0;
             }
@@ -1336,10 +1336,10 @@ bool IDECDROMDevice::doReadCD(uint32_t lba, uint32_t length, uint8_t sector_type
             }
         }
 
-        dbgmsg("---- Read CD: ", (int)length, " sectors starting at ", (int)lba,
+        dbgmsg("---- Read CD: ", (uint64_t)length, " sectors starting at ", (uint64_t)lba,
             ", track number ", trackinfo.track_number, ", sector size ", (int)trackinfo.sector_length,
             ", main channel ", main_channel, ", sub channel ", sub_channel,
-            ", data offset in file ", (int)offset);
+            ", data offset in file ", (uint64_t)offset);
 
         // Ensure read is not out of range of the image
         // If it is, we may be able to get more from the next .bin file.
@@ -1480,7 +1480,7 @@ bool IDECDROMDevice::doReadCD(uint32_t lba, uint32_t length, uint8_t sector_type
         }
         else
         {
-            dbgmsg("-- CD read failed, starting offset ", (int)offset, " length ", (int)length);
+            dbgmsg("-- CD read failed, starting offset ", (uint64_t)offset, " length ", (uint64_t)length);
             return atapi_cmd_error(ATAPI_SENSE_MEDIUM_ERROR, 0);
         }
 
@@ -2137,7 +2137,7 @@ void IDECDROMDevice::atapi_set_mode_page(uint8_t page_ctrl, uint8_t page_idx, co
 bool IDECDROMDevice::doPlayAudio(uint32_t lba, uint32_t length)
 {
 #ifdef ENABLE_AUDIO_OUTPUT
-    dbgmsg("------ CD-ROM Play Audio request at ", (int) lba, " for ",  (int)length, " sectors");
+    dbgmsg("------ CD-ROM Play Audio request at ", (uint64_t) lba, " for ",  (uint64_t)length, " sectors");
 
     // Per Annex C terminate playback immediately if already in progress on
     // the current target. Non-current targets may also get their audio
@@ -2160,8 +2160,8 @@ bool IDECDROMDevice::doPlayAudio(uint32_t lba, uint32_t length)
 
         uint64_t offset = m_cd_read_format.trackinfo.file_offset
                 + trackinfo.sector_length * (lba - trackinfo.track_start);
-        dbgmsg("------ Play audio CD: ", (int)length, " sectors starting at ", (int)lba,
-           ", track number ", trackinfo.track_number, ", data offset in file ", (int)offset);
+        dbgmsg("------ Play audio CD: ", (uint64_t)length, " sectors starting at ", (uint64_t)lba,
+           ", track number ", trackinfo.track_number, ", data offset in file ", (uint64_t)offset);
 
         if (trackinfo.track_mode != CUETrack_AUDIO)
         {
